refactor: index/lazy convention over client/server
