+++ app/[locale]/wallets/find-wallet/page.tsx
@@ -8,7 +8,7 @@ import {
 import { Lang } from "@/lib/types"
 
 import Breadcrumbs from "@/components/Breadcrumbs"
-import FindWalletProductTable from "@/components/FindWalletProductTable/server"
+import FindWalletProductTable from "@/components/FindWalletProductTable/lazy"
 import I18nProvider from "@/components/I18nProvider"
 import MainArticle from "@/components/MainArticle"
