+++ src/components/FindWalletProductTable/lazy.tsx
@@ -1,0 +1,5 @@
+import dynamic from "next/dynamic"
+
+import Loading from "./loading"
+
+export default dynamic(() => import("."), { ssr: false, loading: Loading })
