--- src/components/FindWalletProductTable/server.tsx
@@ -1,8 +1,0 @@
-import dynamic from "next/dynamic"
-
-import Loading from "./loading"
-
-export default dynamic(() => import("./client"), {
-  ssr: false,
-  loading: Loading,
-})
